Ecomerce new/requirements.txt
asgiref==3.3.1
da-vinci==0.3.0
Django==3.1.4
django-admin-thumbnails==0.2.5
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-thumbnails==0.3.2
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Pillow==8.0.1
psycopg2==2.8.6
PyJWT==2.0.1
python-gettext==4.0
pytz==2020.4
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
